feat: Add support for viewing spurious failures
This adds the ability to inspect our spurious failures from today backwards in
time to a given date.

This is done through a simple date-picker in the UI, which feeds data from a
`/spurious/{since_time}` endpoint in the backend, which returns a list of all
the tests, grouped by name, and sorted by the number of failures since the given time.

Future work will include:

* Linking to the tests
* Adding an interval to inspect, both to and from
* Add more test statistics, like which ones have stayed a constant pain, which
are new, and which have likely been fixed, since they stopped showing.

Changelog: None
Ticket: QA-438

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>